refactor: move OptionType to a common type
