+++ src/types/index.ts
@@ -9,10 +9,16 @@ export type Answer = Pick<Pokemon, "name" | "name2">;
 
 export type Selector = [Answer, Answer, Answer, Answer];
 
-export type QuizOption = {
-  isSelectableQuiz: boolean;
-  maxCount: number;
-  selectVersions: SelectVersions;
+export type OptionsType = {
+  numberOfQuiz: number;
+  isChoice: boolean;
+  showHint: boolean;
+  isSilhouette: boolean;
+  hasRegion: boolean;
+  hasAnotherForm: boolean;
+  hasMega: boolean;
+  hasGigantic: boolean;
+  versions: SettingOptions;
 };
 
 export type SettingOptions = {
